QFusionStyle: Fix compilation with qreal = float

+++ src/widgets/styles/qfusionstyle.cpp
@@ -801,8 +801,8 @@ void QFusionStyle::drawPrimitive(PrimitiveElement elem,
 
             } else if (checkbox->state & (State_On)) {
                 qreal penWidth = QStyleHelper::dpiScaled(1.8);
-                penWidth = qMax(penWidth , 0.18 * rect.height());
-                penWidth = qMin(penWidth , 0.30 * rect.height());
+                penWidth = qMax(penWidth , qreal(0.18 * rect.height()));
+                penWidth = qMin(penWidth , qreal(0.30 * rect.height()));
                 QPen checkPen = QPen(checkMarkColor, penWidth);
                 checkMarkColor.setAlpha(210);
                 painter->translate(-0.8, 0.5);
@@ -1584,7 +1584,7 @@ void QFusionStyle::drawControl(ControlElement element, const QStyleOption *optio
             bool ignoreCheckMark = false;
             const int checkColHOffset = windowsItemHMargin + windowsItemFrame - 1;
             int checkcol = qMax(menuItem->rect.height() * 0.7,
-                                      qMax(menuItem->maxIconWidth * 1.0, dpiScaled(17))); // icon checkbox's highlihgt column width
+                                      qMax(qreal(menuItem->maxIconWidth * 1.0), dpiScaled(17))); // icon checkbox's highlihgt column width
             if (
 #if QT_CONFIG(combobox)
                 qobject_cast<const QComboBox*>(widget) ||
